+++ NEWS.md
@@ -1,7 +1,11 @@
 # ospsuite.bmlm NEWS
 
 
-## [0.1.3] - 2024-02-10
+## [0.1.5] - 2024-02-24
+- Add function evaluateAtInitialValues, openLogfile
+- Add filter on outputPathid and scenario to checkfunctions
+
+## [0.1.4] - 2024-02-10
 - Allow flat distribution for individual parameters
 
 ## [0.1.3] - 2024-02-10
@@ -11,7 +15,7 @@
 - Bug fixes in update parameters, only the parameter of the first scenario are updated
 
 ## [0.1.1] - 2024-02-04
-- Bug fixes in chekcProgress functions and export functions 
+- Bug fixes in checkProgress functions and export functions 
   Add publix functions for class BMLMOptimization exportIndividualResultsToPkml, getCurrentConfigTable